Rename apiUtils to events
