+++ src/index.js
@@ -30,7 +30,7 @@ import Dataset from './sources/Dataset';
 import GeoJSON from './sources/GeoJSON';
 import MVT from './sources/MVT';
 import SQL from './sources/SQL';
-import { on, off } from './apiUtils';
+import { on, off } from './utils/events';
 
 /**
  * The version of CARTO VL in use as specified in `package.json` and the GitHub release.
